requirements.txt
streamlit
langchain
openai
python-dotenv
streamlit-chat
duckduckgo-search
tabulate
sentence-transformers
chroma
chromadb
pypdf
tiktoken